An unsigned int is not always big enough to fill a size_t.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@18823 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/HUDNavigationQueue.h
@@ -57,7 +57,7 @@ public:
   void addCallback(HUDNavigationCallback cb, void* data);
   void removeCallback(HUDNavigationCallback cb, void* data);
 
-  static const size_t SkipSetFocus = ~0U;
+  static const size_t SkipSetFocus = ~(size_t)0;
 
 private:
   bool internal_set(size_t index, HUDNavChangeMethod changeMethod, bool setFocus = true);
